Kept developer's packaging for global database
